Add files via upload
Node.js Server that interacts with the IUCN Red List API and Pocket API. Listens on the localhost and port 3000 for incoming requests.
Uses 3LO authentication for sign-in to Pocket; user requires an account.
To use, start the server in a terminal with the command "node index.js".
Then, type localhost:3000 into your browser. Once you hit enter, you will receive a webpage with a search bar. Type the scientific name of an animal (for example, loxodonta africana for the African Elephant) and click search. The application will make the proper requests and add the IUCN webpage for the species to your Pocket List.